+++ writeengine/bulk/we_columnautoinc.h
@@ -16,7 +16,7 @@
    MA 02110-1301, USA. */
 
 /*
-* $Id: we_columnautoinc.h 4450 2013-01-21 14:13:24Z rdempsey $
+* $Id: we_columnautoinc.h 3720 2012-04-04 18:18:49Z rdempsey $
 */
 
 /** @file we_columnautoinc.h
@@ -64,22 +64,22 @@ public:
      * @param autoIncCount Number of auto-increment numbers to reserve
      * @param nextValue Starting number of reserved range
      */
-    virtual int reserveNextRange(uint32_t autoIncCount,
-                        uint64_t& nextValue) = 0;
+    virtual int reserveNextRange(uint autoIncCount,
+                        long long& nextValue) = 0;
 
     /** @brief Finished with auto-incrementing; perform any applicable updates.
      */
     int finish( );
 
 protected:
-    void initNextAutoInc(uint64_t nextValue);
-    uint64_t getNextAutoIncToSave( );
-    int getNextValueFromSysCat(uint64_t& nextValue);
+    void initNextAutoInc(long long nextValue);
+    long long getNextAutoIncToSave( );
+    int getNextValueFromSysCat(long long& nextValue);
 
     Log*        fLog;           // import log file
     boost::mutex fAutoIncMutex; // Mutex to manage fAutoIncLastValue
-    uint64_t    fAutoIncLastValue;// Tracks latest autoincrement value used
-    uint64_t    fMaxIntSat;     // Maximum saturation value
+    long long   fAutoIncLastValue;// Tracks latest autoincrement value used
+    long long   fMaxIntSat;     // Maximum saturation value
     std::string fTableName;     // Full table name (schema.table) for AI column
     std::string fColumnName;    // Name of auto-increment column
     OID         fColumnOID;     // Column OID
@@ -102,8 +102,8 @@ public:
     explicit ColumnAutoIncJob(Log* logger);
     virtual ~ColumnAutoIncJob();
 
-    virtual int reserveNextRange(uint32_t autoIncCount,
-                        uint64_t& nextValue);
+    virtual int reserveNextRange(uint autoIncCount,
+                        long long& nextValue);
 };
 
 //------------------------------------------------------------------------------
@@ -121,8 +121,8 @@ public:
     explicit ColumnAutoIncIncremental(Log* logger);
     virtual ~ColumnAutoIncIncremental();
 
-    virtual int reserveNextRange(uint32_t autoIncCount,
-                        uint64_t& nextValue);
+    virtual int reserveNextRange(uint autoIncCount,
+                        long long& nextValue);
 };
 
 } //end of namespace
